move logic from resolver to service
